requirements.txt
flask==2.3.3
twilio==8.8.0
google-api-python-client==2.97.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
python-dotenv==1.0.0
re
